added another main loop and a new worker that rewrites all topics
Added a new main loop that instead of query per topic we query for all news from the best sources and see if any articles from the best sources are important to what we monitor. Aimed at retreiving more top quality data.

Added a file "worker" or a process that just rewrite all separate analysis/ reports for all topics. To update to the new standards when needed and when changes are done to research process.
also added a file where we can "get all old articles" if one would be running multiple versions of the projects in parallel like i did.