Improved router even more! Also added that we run "new link! only 10% if topic has more than 10 links. Like the other branch
Now the router checks what LLMs are open before using the most likely to be free soon! Really maximizing the local and the cloud servers!

Also added the code from the other branch that makes "new_link" run only 10% of the time if topic has more than 10 links so now we can remove that branch!